Create OOP Calculator with user choice menu - JAVA
This a calculator designed using the OOP in JAVA. This a calculator that provides the user with a menu which enables them to select which operation they would like to perform and then enables them to enter the numbers and perform calculations. It also catches errors and bugs which might occur while the user is giving an input to the system.